% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{assert.equal.length}
\alias{assert.equal.length}
\title{Check for equal length inputs}
\usage{
assert.equal.length(arg0, ...)
}
\arguments{
\item{arg0}{At least one input argument to be checked needs to be given}

\item{...}{Other inputs whose length equality needs to be checked}
}
\value{
Stops execution in case of failed check
}
\description{
Ensure that the given arguments are equal length vectors
}
