% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{assert.for.year}
\alias{assert.for.year}
\title{Check \code{for.year} argument}
\usage{
assert.for.year(for.year)
}
\arguments{
\item{for.year}{input argument to be checked}
}
\value{
Stops execution in case of failed check
}
\description{
Ensure that the for.year argument is a numeric scalar value
}
