% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{assert.numeric.vector}
\alias{assert.numeric.vector}
\title{Check for numeric vector}
\usage{
assert.numeric.vector(arg)
}
\arguments{
\item{arg}{input argument to be checked}
}
\value{
Stops execution in case of failed check
}
\description{
Ensure that the given argument is a numeric vector
}
