\name{etowah}
\alias{etowah}
\docType{data}
\title{ Etowah River Data }
\description{
  A data series from USGS station 02392000, Etowah River (Canton,GA).
}
\usage{
  data(etowah)
}
\format{
  The data are supplied as a matrix with years in the first column and average daily discharge in the second column.
  The series contains 18282 observations starting Jan 1, 1960 and ending Dec 31, 2009.
}
\source{
  \itemize{
    \item{USGS National Water Information Sysem: http://waterdata.usgs.gov/usa/nwis/uv?site_no=02392000}
  }
}


