% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBaseline.R
\name{findMed}
\alias{findMed}
\title{Find median value}
\usage{
findMed(data, indices)
}
\arguments{
\item{data}{complete input vector}

\item{indices}{logical vector to filter indices from \code{data}}
}
\value{
median value of the filtered vector
}
\description{
Find median value from the filtered part of the input vector.
This function is used as a helper to the bootstrapping routine in prepare baseline
}
