% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimingHSAM.R
\name{getTimingHSAM}
\alias{getTimingHSAM}
\title{Time of occurence of High Spectral Anomaly Magnitude (HSAM)}
\usage{
getTimingHSAM(index.hsam, index.ref, years, for.year = NULL)
}
\arguments{
\item{index.hsam}{A scalar/vector of index of HSAM values in given year/years}

\item{index.ref}{A scalar/vector of index of reference point in given year/years}

\item{years}{A vector of years corresponding to HSAM and ref values.
This argument can be NULL if the HSAM and ref values are scalars.}

\item{for.year}{(optional) Calculate timing (HSAM) only for the given year in this argument.
If argument is omitted, timing (HSAM) values for all years are calculated.}
}
\value{
Scalar timing HSAM value if the inputs are scalars, or a Data frame containing two Columns:
\tabular{ll}{
\code{year} \tab First column, represents year \cr
\code{timing.hsam} \tab Second column, represents hsam timing values
}
}
\description{
Compute the number of days separating HSAM and reference point for each year.
}
\examples{
# load sample data
data("sycamore")
x = sycamore

# get streamflow object for the sample data
x.streamflow = asStreamflow(x)

# prepare baseline signal 
x.bl = prepareBaseline(x.streamflow)

# get signal parts
x.sp = getSignalParts(x.bl$pred2, candmin = c(40:125), candmax = c(190:330),
                      years = x.streamflow$data$year, 
                      months = x.streamflow$data$month,
                      jdays = x.streamflow$data$jday)

# get HSAM values
hsam = getHSAM(x.bl$resid.sig, x.streamflow$data$year)

# timing HSAM
thsam = getTimingHSAM(hsam$Index.all, x.sp$peak.index, x.sp$year)

}
