\name{sycamore}
\alias{sycamore}
\docType{data}
\title{ Sycamore Creek Data }
\description{
  A data series from USGS station 09510200, Sycamore Creek (Fort McDowell, AZ).
}
\usage{
  data(sycamore)
}
\format{
  The data are supplied as a matrix with years in the first column and average daily discharge in the second column.
  The series contains 18263 observations starting Jan 1, 1961 and ending Jan 1, 2011.
}
\source{
  \itemize{
    \item{USGS National Water Information Sysem: http://waterdata.usgs.gov/usa/nwis/uv?09510200}
  }
}


